+++ src/ResourceManager/Network/ChangeLog.md
@@ -19,6 +19,44 @@
 -->
 ## Current Release 
 
+## Version 3.5.0
+* Added support for network Watcher APIs
+    - New-AzureRmNetworkWatcher
+    - Get-AzureRmNetworkWatcher
+    - Remove-AzureRmNetworkWatcher
+    - New-AzureRmPacketCaptureFilterConfig
+    - New-AzureRmNetworkWatcherPacketCapture
+    - Get-AzureRmNetworkWatcherPacketCapture
+    - Stop-AzureRmNetworkWatcherPacketCapture
+    - Remove-AzureRmNetworkWatcherPacketCapture
+    - Get-AzureRmNetworkWatcherFlowLogSatus
+    - Get-AzureRmNetworkWatcherNextHop
+    - Get-AzureRmNetworkWatcherSecurityGroupView
+    - Get-AzureRmNetworkWatcherTopology
+    - Get-AzureRmNetworkWatcherTroubleshootingResult
+    - Set-AzureRmNetworkWatcherConfigFlowLog
+    - Start-AzureRmNetworkWatcherResourceTroubleshooting
+    - Test-AzureRmNetworkWatcherIPFlow
+* Add-AzureRmExpressRouteCircuitPeeringConfig
+    - Added new param :-RouteFilter to associate the BGP with route filter to filter out BGP communities via Microsoft peering. This parameter is set by resource
+    - Added new param :-RouteFilterId to associate the BGP with route filter to filter out BGP communities via Microsoft peering. This parameter is set by resource id
+* New-AzureRmExpressRouteCircuitPeeringConfig
+    - Added new param :-RouteFilter to associate the BGP with route filter to filter out BGP communities via Microsoft peering. This parameter is set by resource
+    - Added new param :-RouteFilterId to associate the BGP with route filter to filter out BGP communities via Microsoft peering. This parameter is set by resource id
+* Set-AzureRmExpressRouteCircuitPeeringConfig
+    - Added new param :-RouteFilter to associate the BGP with route filter to filter out BGP communities via Microsoft peering. This parameter is set by resource
+    - Added new param :-RouteFilterId to associate the BGP with route filter to filter out BGP communities via Microsoft peering. This parameter is set by resource id
+* New cmdlets for selective service feature 
+    - Get-AzureRmRouteFilter
+    - New-AzureRmRouteFilter
+    - Set-AzureRmRouteFilter
+    - Remove-AzureRmRouteFilter
+    - Add-AzureRmRouteFilterRuleConfig
+    - Get-AzureRmRouteFilterRuleConfigobject
+    - New-AzureRmRouteFilterRuleConfig
+    - Set-AzureRmRouteFilterRuleConfig
+    - Remove-AzureRmRouteFilterRuleConfig
+
 ## Version 3.4.0
 
 ## Version 3.3.0
@@ -37,3 +75,4 @@
 * Get-AzureRmEffectiveNetworkSecurityGroup
     - Add warning if there is no response from GetEffectiveNSG
 * Add Source property to EffectiveRoute 
+